requirements.txt
ansible>=2.7
molecule>=3.0.0
docker
requests
molecule-vagrant
python-vagrant
pytest
pytest-testinfra==6.1.0